Fighter is now stable. Replacing AbstractFighter
